+++ src/main/java/ua/org/ubts/applicationssystem/service/ProgramService.java
@@ -4,7 +4,7 @@ import ua.org.ubts.applicationssystem.entity.Program;
 import ua.org.ubts.applicationssystem.service.template.BasicService;
 import ua.org.ubts.applicationssystem.service.template.ExtendedElementOperationsService;
 
-public interface ProgramService extends BasicService<Program, Long>, ExtendedElementOperationsService<Program> {
+public interface ProgramService extends BasicService<Program, Integer>, ExtendedElementOperationsService<Program> {
 
     Program findByNameAndInfo(String name, String info);
