+++ src/main/java/ua/org/ubts/applicationssystem/service/ResidenceService.java
@@ -3,6 +3,5 @@ package ua.org.ubts.applicationssystem.service;
 import ua.org.ubts.applicationssystem.entity.Residence;
 import ua.org.ubts.applicationssystem.service.template.BasicService;
 
-public interface ResidenceService extends BasicService<Residence, Long> {
-
+public interface ResidenceService extends BasicService<Residence, Integer> {
 }
