+++ src/main/java/ua/org/ubts/applicationssystem/service/StudentService.java
@@ -7,7 +7,7 @@ import ua.org.ubts.applicationssystem.service.template.ExtendedElementOperations
 /**
  * Created by Yaroslav on 18.07.2017.
  */
-public interface StudentService extends BasicService<Student, Long>, ExtendedElementOperationsService<Student> {
+public interface StudentService extends BasicService<Student, Integer>, ExtendedElementOperationsService<Student> {
 
     Student findByName(String firstName, String middleName, String lastName);
